Update Pandoc installation in SDK generation workflow
Modified the Pandoc installation step in the '.github/workflows/generate-sdk.yaml' file to use the 'lastversion' command instead of directly installing '*.deb' files. This ensures that the latest version of Pandoc, suitable for the system architecture, is always used. This improvement is vital for keeping our SDK documentation generation process updated.