Use python module for docformatter in Github workflow
The command in the 'Format Code' step of our Github workflow has been updated to call the docformatter package as a python module. This change rectifies an issue where the docformatter couldn't be found when invoked via the sudo command, ensuring smooth execution of our code formatting in the future.