"Allow lint workflow to trigger on any push
Extended the triggering conditions for 'Do Lint' and 'Quality Checks' workflows in GitHub Actions. Instead of specifying branches, the condition now applies to any push made in any branch of the repository. This change improves automation, cuts back on manual triggers and allows the Lint process to run irrespective of the branches used."