"Refactor lint triggering mechanism in workflows
Changes have been made to the triggering conditions in 'Do Lint' and 'Quality Checks' yaml workflows. Earlier, execution was dependent on the completion of other workflows and specific to branch names. This has now been replaced with a repository dispatch event, allowing the workflows to trigger on any push to any branch. This improves process efficiency and automation, ensuring lint checks are performed for all push events, irrespective of the branches in use."