Refactor GitHub Actions and improve code format management.
To streamline the CI/CD pipeline, the redundant 'run-formatters' job was removed from the 'quality.yaml' workflow as linting tasks were parameterized and moved to the 'generator-publish-sources.yaml' workflow. This was done to ensure that code formatting checks and modifications are performed while preparing the release source files, guaranteeing the code quality in the release branch without the need for a separate job.