Update quality.yaml to trigger on both push and PR
Change the trigger for the quality checks in Github Actions to run both on 'push' and 'pull_request' events. This ensures that the quality checks are ran not only when changes are pushed to the main branch, but also when a pull request is made to the repo, increasing the overall codebase's stability and reliability.